Consolidated intent and optmization functionality
