+++ data/output/enriched_products.json
@@ -18,8 +18,7 @@
       "dress_shopping",
       "summer"
     ],
-    "ai_optimized_title": "H&M Women Ladies Dress",
-    "ai_optimized_description": "summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton"
+    "ai_optimized_content": "H&M Women Ladies Dress. summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton"
   },
   {
     "product_id": "HM002",
@@ -41,8 +40,7 @@
       "casual",
       "summer"
     ],
-    "ai_optimized_title": "H&M Men Tshirt Men",
-    "ai_optimized_description": "basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit"
+    "ai_optimized_content": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit"
   },
   {
     "product_id": "HM003",
@@ -65,8 +63,7 @@
       "eco_friendly",
       "summer"
     ],
-    "ai_optimized_title": "Eco-Friendly H&M Organic Hoodie",
-    "ai_optimized_description": "hoodie warm soft eco cotton. Perfect for comfort, cozy wear. Features: cotton, organic"
+    "ai_optimized_content": "Eco-Friendly H&M Organic Hoodie. hoodie warm soft eco cotton. Perfect for comfort, cozy wear. Features: cotton, organic"
   },
   {
     "product_id": "HM004",
@@ -88,8 +85,7 @@
       "budget_friendly",
       "comfort"
     ],
-    "ai_optimized_title": "H&M Kids Kids Jeans Blue",
-    "ai_optimized_description": "stretch denim pants kids. Perfect for budget friendly, comfort. Features: denim, stretchy, blue color"
+    "ai_optimized_content": "H&M Kids Kids Jeans Blue. stretch denim pants kids. Perfect for budget friendly, comfort. Features: denim, stretchy, blue color"
   },
   {
     "product_id": "HM005",
@@ -108,8 +104,7 @@
     "intents": [
       "casual"
     ],
-    "ai_optimized_title": "H&M Sneaker Shoes",
-    "ai_optimized_description": "white shoes unisex everyday. Perfect for casual. Features: white color"
+    "ai_optimized_content": "H&M Sneaker Shoes. white shoes unisex everyday. Perfect for casual. Features: white color"
   },
   {
     "product_id": "HM006",
@@ -126,7 +121,6 @@
     "intents": [
       "comfort"
     ],
-    "ai_optimized_title": "H&M Winter Coat",
-    "ai_optimized_description": "warm waterproof coat. Perfect for comfort"
+    "ai_optimized_content": "H&M Winter Coat. warm waterproof coat. Perfect for comfort"
   }
 ]
