+++ data/output/knowledge_graph.json
@@ -1,7 +1,7 @@
 {
   "products": {
     "HM001": {
-      "title": "H&M Women Ladies Dress",
+      "description": "H&M Women Ladies Dress. summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton",
       "category": "clothes/women/dresses",
       "intents": [
         "budget_friendly",
@@ -14,7 +14,7 @@
       "price": 29.99
     },
     "HM002": {
-      "title": "H&M Men Tshirt Men",
+      "description": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit",
       "category": "mens/tops",
       "intents": [
         "budget_friendly",
@@ -28,7 +28,7 @@
       "price": 12.99
     },
     "HM003": {
-      "title": "Eco-Friendly H&M Organic Hoodie",
+      "description": "Eco-Friendly H&M Organic Hoodie. hoodie warm soft eco cotton. Perfect for comfort, cozy wear. Features: cotton, organic",
       "category": "unisex/hoodies",
       "intents": [
         "comfort",
