+++ data/output/query_results.json
@@ -2,100 +2,100 @@
   "affordable summer dresses under $30": [
     {
       "product_id": "HM001",
-      "title": "H&M Women Ladies Dress",
-      "score": 0.8193364024162293,
+      "description": "H&M Women Ladies Dress. summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton",
+      "score": 0.8080576539039612,
       "reason": "Strong semantic match. Price in range ($29.99)"
     },
     {
       "product_id": "HM002",
-      "title": "H&M Men Tshirt Men",
-      "score": 0.634555208683014,
+      "description": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit",
+      "score": 0.6232020139694214,
       "reason": "Price in range ($12.99)"
     },
     {
       "product_id": "HM004",
-      "title": "H&M Kids Kids Jeans Blue",
-      "score": 0.5029468655586242,
+      "description": "H&M Kids Kids Jeans Blue. stretch denim pants kids. Perfect for budget friendly, comfort. Features: denim, stretchy, blue color",
+      "score": 0.5088505268096923,
       "reason": "Price in range ($19.99)"
     }
   ],
   "eco-friendly hoodies for everyday wear": [
     {
       "product_id": "HM003",
-      "title": "Eco-Friendly H&M Organic Hoodie",
-      "score": 0.687059760093689,
+      "description": "Eco-Friendly H&M Organic Hoodie. hoodie warm soft eco cotton. Perfect for comfort, cozy wear. Features: cotton, organic",
+      "score": 0.682920515537262,
       "reason": "Strong semantic match"
     },
     {
       "product_id": "HM002",
-      "title": "H&M Men Tshirt Men",
-      "score": 0.36660879850387573,
+      "description": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit",
+      "score": 0.3621394634246826,
       "reason": "Partial match"
     },
     {
       "product_id": "HM001",
-      "title": "H&M Women Ladies Dress",
-      "score": 0.3322712779045105,
+      "description": "H&M Women Ladies Dress. summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton",
+      "score": 0.3306998014450073,
       "reason": "Partial match"
     }
   ],
   "basic t-shirts for men slim fit": [
     {
       "product_id": "HM002",
-      "title": "H&M Men Tshirt Men",
-      "score": 0.7953287839889527,
+      "description": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit",
+      "score": 0.787356722354889,
       "reason": "Strong semantic match. Has slim fit"
     },
     {
-      "product_id": "HM001",
-      "title": "H&M Women Ladies Dress",
-      "score": 0.33122357726097107,
+      "product_id": "HM005",
+      "description": "H&M Sneaker Shoes. white shoes unisex everyday. Perfect for casual. Features: white color",
+      "score": 0.3489392101764679,
       "reason": "Partial match"
     },
     {
-      "product_id": "HM005",
-      "title": "H&M Sneaker Shoes",
-      "score": 0.3278101682662964,
+      "product_id": "HM001",
+      "description": "H&M Women Ladies Dress. summer cotton dress. Perfect for budget friendly, dress shopping. Features: cotton",
+      "score": 0.3380645215511322,
       "reason": "Partial match"
     }
   ],
   "comfortable jeans for kids": [
     {
       "product_id": "HM004",
-      "title": "H&M Kids Kids Jeans Blue",
-      "score": 0.7735329031944275,
+      "description": "H&M Kids Kids Jeans Blue. stretch denim pants kids. Perfect for budget friendly, comfort. Features: denim, stretchy, blue color",
+      "score": 0.762129282951355,
       "reason": "Strong semantic match"
     },
     {
       "product_id": "HM003",
-      "title": "Eco-Friendly H&M Organic Hoodie",
-      "score": 0.4074846863746643,
+      "description": "Eco-Friendly H&M Organic Hoodie. hoodie warm soft eco cotton. Perfect for comfort, cozy wear. Features: cotton, organic",
+      "score": 0.4070111870765686,
       "reason": "Partial match"
     },
     {
       "product_id": "HM006",
-      "title": "H&M Winter Coat",
-      "score": 0.3490536570549011,
+      "description": "H&M Winter Coat. warm waterproof coat. Perfect for comfort",
+      "score": 0.3495497316122055,
       "reason": "Partial match"
     }
   ],
   "white sneakers for casual outfits": [
     {
       "product_id": "HM005",
-      "title": "H&M Sneaker Shoes",
-      "score": 0.8154737710952759,
+      "description": "H&M Sneaker Shoes. white shoes unisex everyday. Perfect for casual. Features: white color",
+      "score": 0.8352731227874756,
       "reason": "Strong semantic match"
     },
     {
       "product_id": "HM002",
-      "title": "H&M Men Tshirt Men",
-      "score": 0.4443295657634735,
+      "description": "H&M Men Tshirt Men. basic tee cotton slim fit. Perfect for budget friendly, casual. Features: cotton, slim fit",
+      "score": 0.4380874991416931,
       "reason": "Partial match"
     },
     {
       "product_id": "HM004",
-      "title": "H&M Kids Kids Jeans Blue",
-      "score": 0.3063380718231201,
+      "description": "H&M Kids Kids Jeans Blue. stretch denim pants kids. Perfect for budget friendly, comfort. Features: denim, stretchy, blue color",
+      "score": 0.31391197443008423,
       "reason": "Partial match"
     }
   ]
